pyproject.toml
[project]
name = "mcp-notebooklm"
version = "0.1.0"
description = "MCP Server for Google NotebookLM"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "notebooklm-py[playwright]>=0.3.0",
    "python-dotenv>=1.0.1",
    "playwright>=1.40.0",
]

[project.scripts]
mcp-notebooklm = "mcp_notebooklm.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
